Refs #23919 -- Removed unneeded float()/int() calls.

+++ django/db/models/fields/__init__.py
@@ -1631,7 +1631,7 @@ class DurationField(Field):
         if value is None:
             return None
         # Discard any fractional microseconds due to floating point arithmetic.
-        return int(round(value.total_seconds() * 1000000))
+        return round(value.total_seconds() * 1000000)
 
     def get_db_converters(self, connection):
         converters = []
